changed data type from int to double in Calculator class, added multiply and divide test

+++ src/main/java/edward_dunn/unit_test_examples/Calculator.java
@@ -8,19 +8,19 @@ public class Calculator {
 		
 	}
 	
-	public int add(int value1, int value2) {
+	public double add(double value1, double value2) {
 		return value1 + value2;
 	}
 	
-	public int subtract(int value1, int value2) {
+	public double subtract(double value1, double value2) {
 		return value1 - value2;
 	}
 	
-	public int multiply(int value1, int value2) {
+	public double multiply(double value1, double value2) {
 		return value1 * value2;
 	}
 	
-	public int divide(int value1, int value2) {
+	public double divide(double value1, double value2) {
 		return value1 / value2;
 	}
 }
